+++ inst/examples/getReactive.r
@@ -1,7 +1,7 @@
 \dontrun{
 
 where <- new.env()  
-setThis(id = "x_1", value = TRUE, where = where)
+setReactive(id = "x_1", value = TRUE, where = where)
 getThis(id = "x_1", where = where)
 
 }