+++ rapp/dev/example.r
@@ -1,18 +1,18 @@
 x_1 <- new.env()  
 x_2 <- new.env()  
   
-setThis(id = "field_1", value = Sys.time(), envir = x_1)
+setReactive(id = "field_1", value = Sys.time(), envir = x_1)
 getThis(id = "x_1", envir = envir)
 
 binding <- substitute(function(x) {
   x + 60*60*24
 })
-setThis(id = "x_2", envir = envir, binding = binding, watch = "x_1", 
+setReactive(id = "x_2", envir = envir, binding = binding, watch = "x_1", 
          binding_type = 2)
 getThis(id = "x_2", envir = envir)  
   
 ## Change value of monitored variable //
-setThis(id = "x_1", value = Sys.time(), envir = envir, binding_type = 2)
+setReactive(id = "x_1", value = Sys.time(), envir = envir, binding_type = 2)
 getThis(id = "x_1", envir = envir)  
 getThis(id = "x_2", envir = envir)    
  
@@ -40,7 +40,7 @@ b1 <- substitute(
   list(VALUE = as.name("value"))
 )
 
-setThis(
+setReactive(
   id = "x_1", 
   value = 10, 
   envir = envir, 
@@ -95,7 +95,7 @@ b2 <- substitute(
   list(VALUE = as.name("value"))
 )
 
-setThis(
+setReactive(
   id = "x_2", 
   envir = envir, 
   watch = "x_1", 
@@ -111,7 +111,7 @@ envir$x_1
 envir$x_2
 envir$x_2
 
-setThis(
+setReactive(
   id = "x_1", 
   value = 100, 
   envir = envir, 
@@ -129,14 +129,14 @@ require("microbenchmark")
 
 envir <- new.env()  
 res_1 <- microbenchmark(
-  "1" = setThis(id = "x_1", value = Sys.time(), envir = envir, binding_type = 2),
+  "1" = setReactive(id = "x_1", value = Sys.time(), envir = envir, binding_type = 2),
   "2" = getThis(id = "x_1", envir = envir),
-  "3" = setThis(id = "x_2", envir = envir,
+  "3" = setReactive(id = "x_2", envir = envir,
     binding = substitute(function(x) {
         x + 60*60*24
       }), watch = "x_1", binding_type = 2),
   "4" = getThis(id = "x_2", envir = envir),
-  "5" = setThis(id = "x_1", value = Sys.time(), envir = envir,
+  "5" = setReactive(id = "x_1", value = Sys.time(), envir = envir,
                  binding_type = 2),
   "6" = getThis(id = "x_2", envir = envir),
   control = list(order = "inorder", warmup = 10)
@@ -147,12 +147,12 @@ res_1
 
 envir <- new.env()
 res_2 <- microbenchmark(
-  "1" = setThis(id = "x_1", value = 10, envir = envir, binding = b1,
+  "1" = setReactive(id = "x_1", value = 10, envir = envir, binding = b1,
     binding_type = 1),
   "2" = getThis(id = "x_1", envir = envir),
-  "3" = setThis(id = "x_2", envir = envir, binding = b2, watch = "x_1"),
+  "3" = setReactive(id = "x_2", envir = envir, binding = b2, watch = "x_1"),
   "4" = getThis(id = "x_2", envir = envir),
-  "5" = setThis(id = "x_1", value = 100, envir = envir),
+  "5" = setReactive(id = "x_1", value = 100, envir = envir),
   "6" = getThis(id = "x_2", envir = envir),
   control = list(order = "inorder", warmup = 10)
 )