+++ tests/testthat/test-getReactive.r
@@ -3,7 +3,7 @@ test_that("getThis", {
 
   ## Create example content //
   where <- new.env()
-  setThis(id = "test", value = new.env(), where = where, 
+  setReactive(id = "test", value = new.env(), where = where, 
            binding_type = 2)
   expect_is(getThis(id = "test", where = where), "environment")