+++ tests/testthat/test-setReactive_bare.r
@@ -1,5 +1,5 @@
-context("setThis-1")
-test_that("setThis", {
+context("setReactive_bare-1")
+test_that("setReactive_bare", {
   
   .hash_id <- "._HASH"
   where = new.env()  
@@ -7,7 +7,7 @@ test_that("setThis", {
   id <- "x_1"
   value <- Sys.time()
   expect_equal(
-    setThis(id = "x_1", value = value, where = where, binding_type = 2),
+    setReactive_bare(id = "x_1", value = value, where = where, binding_type = 2),
     value
   )
   
@@ -26,7 +26,7 @@ test_that("setThis", {
   watch <- "x_1"
   expected <- eval(binding)(x = value)
   expect_equal(
-    setThis(id = id, where = where, binding = binding, watch = watch,
+    setReactive_bare(id = id, where = where, binding = binding, watch = watch,
              binding_type = 2),
     expected
   )
@@ -46,7 +46,7 @@ test_that("setThis", {
   ## Change watch value //
   value = Sys.time()
   expect_equal(
-    setThis(id = "x_1", value = value, where = where, binding_type = 2),
+    setReactive_bare(id = "x_1", value = value, where = where, binding_type = 2),
     value
   )
   expect_true(where[[.hash_id]]$x_1$x_1 != where[[.hash_id]]$x_1$x_2)
@@ -66,7 +66,7 @@ test_that("setThis", {
   value <- 10
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
       value = value, 
       where = where, 
@@ -86,7 +86,7 @@ test_that("setThis", {
   ## Set again //
   value <- 10
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
       value = value, 
       where = where, 
@@ -101,7 +101,7 @@ test_that("setThis", {
   binding <- function(x) {x + 100}
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       where = where, 
       watch = "x_1", 
@@ -134,7 +134,7 @@ test_that("setThis", {
   value <- 500
   expected <- value
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
       value = value, 
       where = where, 
@@ -148,7 +148,7 @@ test_that("setThis", {
   expected <- binding(x = where$x_1)
 
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       where = where, 
       watch = "x_1",
@@ -171,15 +171,15 @@ test_that("setThis", {
     
     where <- new.env()  
     res_1 <- microbenchmark(
-      "1" = setThis(id = "x_1", value = Sys.time(), where = where,
+      "1" = setReactive_bare(id = "x_1", value = Sys.time(), where = where,
                      binding_type = 2),
       "2" = getThis(id = "x_1", where = where),
-      "3" = setThis(id = "x_2", where = where,
+      "3" = setReactive_bare(id = "x_2", where = where,
         binding = substitute(function(x) {
             x + 60*60*24
           }), watch = "x_1", binding_type = 2),
       "4" = getThis(id = "x_2", where = where),
-      "5" = setThis(id = "x_1", value = Sys.time(), where = where,
+      "5" = setReactive_bare(id = "x_1", value = Sys.time(), where = where,
                      binding_type = 2),
       "6" = getThis(id = "x_2", where = where),
       control = list(order = "inorder")
@@ -191,56 +191,31 @@ test_that("setThis", {
     where <- new.env()
     
     res_2 <- microbenchmark(
-      "1" = setThis(id = "x_1", value = 10, where = where),
+      "1" = setReactive_bare(id = "x_1", value = 10, where = where),
       "2" = getThis(id = "x_1", where = where),
-      "3" = setThis(id = "x_2", where = where, watch = "x_1",
+      "3" = setReactive_bare(id = "x_2", where = where, watch = "x_1",
         binding = function(x) {x + 100}),
       "4" = getThis(id = "x_2", where = where),
-      "5" = setThis(id = "x_1", value = 100, where = where),
+      "5" = setReactive_bare(id = "x_1", value = 100, where = where),
       "6" = getThis(id = "x_2", where = where),
       control = list(order = "inorder")
     )
     res_2
-
-     #-----------------
-
-    foo <- function(x) {
-      x + 10
-    }
-    setGeneric(name = "bar", signature=c("x"), 
-               def = function(x) standardGeneric)
-    setMethod(f = "bar", signature = c(x = "ANY"), definition = function(x) {
-      x + 10
-    })
-    
-    require("compiler")
-    setGeneric(name = "foobar", signature=c("x"), 
-               def = cmpfun(function(x) standardGeneric))
-    setMethod(f = "foobar", signature = c(x = "ANY"), 
-              definition = cmpfun(function(x) {
-      x + 10
-    }))
-    res_3 <- microbenchmark(
-      "foo" = foo(x = 10),
-      "bar" = bar(x = 10),
-      "foobar" = foobar(x = 10)
-    )
-    res_3
   }
   
 })
 
 ##------------------------------------------------------------------------------
 
-context("setThis-2")
-test_that("setThis", {
+context("setReactive_bare-2")
+test_that("setReactive_bare", {
   
   where = new.env()  
   id = "x_1"
   value = 10
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = id, 
       value = value, 
       where = where
@@ -248,7 +223,7 @@ test_that("setThis", {
     value
   )
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = id, 
       value = value, 
       where = where
@@ -265,7 +240,7 @@ test_that("setThis", {
 
   expected <- binding(value)
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       where = where, 
       watch = "x_1", 
@@ -274,46 +249,57 @@ test_that("setThis", {
     ),
     expected
   )
-  where$x_1  
-  where$x_2
+  
+  expect_equal(where$x_1, value)
+  expect_equal(where$x_2, binding(where$x_1))
   where$x_1 <- 100
-  where$x_2
+  expect_equal(where$x_2, binding(where$x_1))
 
 })
 
 ##------------------------------------------------------------------------------
 
-context("setThis-3")
-test_that("setThis", {
+context("setReactive_bare-3")
+test_that("setReactive_bare", {
   
   .hash_id <- "._HASH"
   where <- new.env()  
-    
+  value <- 10 
+  
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
-      value = 10, 
+      value = value,
       where = where
     ),
-    10
+    value
   )
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       where = where,
       watch = "x_1"
     ),
-    10
+    value
   )
-  ## Values are ignored if variable monitors another //
+  
+  expect_equal(where$x_1, value)
+  expect_equal(where$x_2, value)
+  
+  where$x_1 <- 100
+  expect_equal(where$x_1, 100)
+  expect_equal(where$x_2, 100)
+  
+  ## Values are ignored if variable monitors another
   expect_warning(
-    setThis(
+    setReactive_bare(
       id = "x_3", 
       value = 20, 
       where = where,
       watch = "x_1"
     )
   )
+  expect_equal(where$x_3, where$x_1)
   
   ##----------------------------------------------------------------------------
   ## Mutual dependency //
@@ -322,7 +308,7 @@ test_that("setThis", {
   where <- new.env()  
     
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
       where = where,
       watch = "x_2",
@@ -333,7 +319,7 @@ test_that("setThis", {
   )
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       where = where,
       watch = "x_1",
@@ -351,7 +337,7 @@ test_that("setThis", {
   expect_equal(where[[.hash_id]]$x_2$x_1, expected)
   expect_equal(where[[.hash_id]]$x_2$x_2, expected)
   
-  ## Initial values //
+  ## Forced initial values //
   expect_equal(where$x_1, NULL)
   expect_equal(where$x_2, NULL)
   
@@ -392,24 +378,24 @@ test_that("setThis", {
 
 ##------------------------------------------------------------------------------
 
-context("setThis-4")
-test_that("setThis", {
+context("setReactive_bare-4")
+test_that("setReactive_bare", {
 
   if (FALSE) {
-    
+  
   suppressWarnings(rm(x_1))
   suppressWarnings(rm(x_2))
   suppressWarnings(rm(.hash))
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
       value = 10
     ),
     10
   )
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       watch = "x_1"
     ),
@@ -417,7 +403,7 @@ test_that("setThis", {
   )
   ## Values are ignored if variable monitors another //
   expect_warning(
-    setThis(
+    setReactive_bare(
       id = "x_3", 
       value = 20, 
       watch = "x_1"
@@ -432,7 +418,7 @@ test_that("setThis", {
   ##----------------------------------------------------------------------------
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_1", 
       watch = "x_2",
       mutual = TRUE,
@@ -442,7 +428,7 @@ test_that("setThis", {
   )
   
   expect_equal(
-    setThis(
+    setReactive_bare(
       id = "x_2", 
       watch = "x_1",
       mutual = TRUE,
@@ -502,63 +488,4 @@ test_that("setThis", {
 
   }
 
-})
-
-##------------------------------------------------------------------------------
-
-context("setThis-5")
-test_that("setThis", {
-
-  tracelevel <- 0
-  where <- new.env()
-  
-  setThis(id = "x_1", value = 10, where = where)
-  setThis(id = "x_2", watch = "x_1", where = where)
-  
-  where$x_1 
-  where$x_2
-  where$x_1 <- 100 
-  where$x_2
-  
-  setThis(id = "x_3", watch = "x_1", where = where, 
-    binding = function(x) {x * 2})
-  
-  where$x_1 
-  where$x_2
-  where$x_3
-  where$x_1 <- 500
-  where$x_2
-  where$x_3  
-  
-  setThis(id = "x_1", watch = "x_4", where = where, mutual = TRUE, 
-          .tracelevel = tracelevel)
-  setThis(id = "x_4", watch = "x_1", where = where, mutual = TRUE,
-          .tracelevel = tracelevel)
-
-  where$x_1
-  where$x_4
-  
-  if (FALSE) {
-    hash_id <- "._HASH"
-    where[[hash_id]]$x_1$x_1
-    where[[hash_id]]$x_1$x_4
-    where[[hash_id]]$x_4$x_4
-    where[[hash_id]]$x_4$x_1
-  }
-  
-  where$x_2
-  where$x_3
-  
-  where$x_1 <- 100
-  where$x_1
-  where$x_4
-  where$x_2
-  where$x_3
-  
-  where$x_4 <- 1000
-  where$x_4
-  expect_equal(where$x_1, 1000)
-  expect_equal(where$x_2, 1000)
-  expect_equal(where$x_3, 2000)
-  
 })